eliminate gerund from cairo description
Description stated aim to foster collaboration, broken down this was
*collaborate in conducting new research* and *collaborate in start-ups*,
but since one was a gerund phrase and the other not, it was correct
while being odd. For balance, one would need to clarify what was
being done with start-ups (add a gerund), or condense the language
about research (since "conducting" is implied, and even studying
new research is itself a form of conducting research).